chore: move usercard to molecules
