+++ app/src/components/molecules/UserCard/index.jsx
@@ -2,9 +2,9 @@ import React from 'react'
 import { Image, Text, View } from 'react-native'
 import { Icon } from 'react-native-elements'
 import { TouchableOpacity } from 'react-native-gesture-handler'
-import callService from '../../services/callService'
+import callService from '../../../services/callService'
 import { styles } from './styles'
-import SessionService from '../../services/Session';
+import SessionService from '../../../services/Session';
 
 export const UserCard = ({ name, email, photo }) => {
     const handleLogout = async () => {
